requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
opencv-python==4.11.0.86
pillow==11.2.1
Werkzeug==3.1.3
